photon voice appid field, generic input scanner
